Fix merge conflicts in PR #184
- Refactored `SubmitPage` to be a Server Component for improved SEO (canonical tags) and layout handling.
- Extracted `SubmitForm` into a Client Component to handle form logic.
- Preserved `pricing_model` and `price` fields in the form.
- Resolved conflicts by manually applying the desired changes from PR #184 while keeping the existing functionality from `main`.
