+++ npm_install.log
@@ -1,0 +1,9 @@
+npm warn deprecated node-domexception@1.0.0: Use your platform's native DOMException instead
+npm warn deprecated glob@10.5.0: Old versions of glob are not supported, and contain widely publicized security vulnerabilities, which have been fixed in the current version. Please update. Support for old versions may be purchased (at exorbitant rates) by contacting i@izs.me
+
+added 574 packages, and audited 575 packages in 32s
+
+262 packages are looking for funding
+  run `npm fund` for details
+
+found 0 vulnerabilities
