+++ src/components/SubmitForm.tsx
@@ -1,0 +1,224 @@
+'use client';
+
+import { useTranslations } from 'next-intl';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { z } from 'zod';
+import { useState } from 'react';
+import { Loader2 } from 'lucide-react';
+
+const submissionSchema = z.object({
+  name: z.string().min(1, 'Tool name is required'),
+  url: z.string().url('Please enter a valid URL'),
+  description: z.string().min(10, 'Description must be at least 10 characters'),
+  category: z.string().min(1, 'Category is required'),
+  pricing_model: z.string().min(1, 'Pricing model is required'),
+  price: z.string().optional(),
+});
+
+type SubmissionFormData = z.infer<typeof submissionSchema>;
+
+export function SubmitForm() {
+  const t = useTranslations('SubmitPage');
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const [submitStatus, setSubmitStatus] = useState<'idle' | 'success' | 'error'>('idle');
+
+  const {
+    register,
+    handleSubmit,
+    reset,
+    formState: { errors },
+  } = useForm<SubmissionFormData>({
+    resolver: zodResolver(submissionSchema),
+  });
+
+  const onSubmit = async (data: SubmissionFormData) => {
+    setIsSubmitting(true);
+    setSubmitStatus('idle');
+
+    try {
+      const response = await fetch('/api/submit', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(data),
+      });
+
+      if (!response.ok) {
+        throw new Error('Submission failed');
+      }
+
+      setSubmitStatus('success');
+      reset();
+    } catch (error) {
+      console.error('Submission error:', error);
+      setSubmitStatus('error');
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
+
+  return (
+    <form onSubmit={handleSubmit(onSubmit)} className="space-y-6">
+      <div>
+        <label htmlFor="name" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.name')}
+        </label>
+        <div className="mt-2">
+          <input
+            id="name"
+            type="text"
+            {...register('name')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 placeholder:text-zinc-400 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+          />
+          {errors.name && (
+            <p className="mt-1 text-sm text-red-600">{errors.name.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="url" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.url')}
+        </label>
+        <div className="mt-2">
+          <input
+            id="url"
+            type="url"
+            {...register('url')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 placeholder:text-zinc-400 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+            placeholder="https://example.com"
+          />
+          {errors.url && (
+            <p className="mt-1 text-sm text-red-600">{errors.url.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="category" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.category')}
+        </label>
+        <div className="mt-2">
+          <input
+            id="category"
+            type="text"
+            {...register('category')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 placeholder:text-zinc-400 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+          />
+          {errors.category && (
+            <p className="mt-1 text-sm text-red-600">{errors.category.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="pricing_model" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.pricing_model')}
+        </label>
+        <div className="mt-2">
+          <select
+            id="pricing_model"
+            {...register('pricing_model')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+          >
+            <option value="">Select a pricing model</option>
+            <option value="free">{t('form.pricing_model_options.free')}</option>
+            <option value="freemium">{t('form.pricing_model_options.freemium')}</option>
+            <option value="paid">{t('form.pricing_model_options.paid')}</option>
+            <option value="contact">{t('form.pricing_model_options.contact')}</option>
+          </select>
+          {errors.pricing_model && (
+            <p className="mt-1 text-sm text-red-600">{errors.pricing_model.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="price" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.price')}
+        </label>
+        <div className="mt-2">
+          <input
+            id="price"
+            type="text"
+            {...register('price')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 placeholder:text-zinc-400 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+          />
+          {errors.price && (
+            <p className="mt-1 text-sm text-red-600">{errors.price.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <label htmlFor="description" className="block text-sm font-medium leading-6 text-zinc-900 dark:text-white">
+          {t('form.description')}
+        </label>
+        <div className="mt-2">
+          <textarea
+            id="description"
+            rows={4}
+            {...register('description')}
+            className="block w-full rounded-md border-0 py-1.5 text-zinc-900 shadow-sm ring-1 ring-inset ring-zinc-300 placeholder:text-zinc-400 focus:ring-2 focus:ring-inset focus:ring-blue-600 sm:text-sm sm:leading-6 dark:bg-zinc-800 dark:text-white dark:ring-zinc-700"
+          />
+          {errors.description && (
+            <p className="mt-1 text-sm text-red-600">{errors.description.message}</p>
+          )}
+        </div>
+      </div>
+
+      <div>
+        <button
+          type="submit"
+          disabled={isSubmitting}
+          className="flex w-full justify-center rounded-md bg-blue-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          {isSubmitting ? (
+            <>
+              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+              {t('form.submitting')}
+            </>
+          ) : (
+            t('form.submit')
+          )}
+        </button>
+      </div>
+
+      {submitStatus === 'success' && (
+        <div className="rounded-md bg-green-50 p-4 dark:bg-green-900/30">
+          <div className="flex">
+            <div className="flex-shrink-0">
+              <svg className="h-5 w-5 text-green-400" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
+                <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.483 4.79-1.88-1.88a.75.75 0 10-1.06 1.061l2.5 2.5a.75.75 0 001.137-.089l4-5.5z" clipRule="evenodd" />
+              </svg>
+            </div>
+            <div className="ml-3">
+              <p className="text-sm font-medium text-green-800 dark:text-green-200">
+                {t('form.success')}
+              </p>
+            </div>
+          </div>
+        </div>
+      )}
+
+      {submitStatus === 'error' && (
+        <div className="rounded-md bg-red-50 p-4 dark:bg-red-900/30">
+          <div className="flex">
+            <div className="flex-shrink-0">
+               <svg className="h-5 w-5 text-red-400" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
+                <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.28 7.22a.75.75 0 00-1.06 1.06L8.94 10l-1.72 1.72a.75.75 0 101.06 1.06L10 11.06l1.72 1.72a.75.75 0 101.06-1.06L11.06 10l1.72-1.72a.75.75 0 00-1.06-1.06L10 8.94 8.28 7.22z" clipRule="evenodd" />
+              </svg>
+            </div>
+            <div className="ml-3">
+              <p className="text-sm font-medium text-red-800 dark:text-red-200">
+                {t('form.error')}
+              </p>
+            </div>
+          </div>
+        </div>
+      )}
+    </form>
+  );
+}
